legal services

+++ app/Http/Controllers/paragraphController.php
@@ -2,8 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-use App\Http\Requests;
 use App\Http\Controllers\Controller;
 use App\Paragraph;
 use App\Paragraph_attrs;
@@ -23,22 +21,18 @@ class paragraphController extends Controller {
         return view('index', $data);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return Response
+    /*
+     * Method to store all kind paragraphs
+     * TYPE can be
+     * P -  pargraphs main page /index
+     * PL -  paragraphs legal page
+     * R - pargraphs index page seeding from seeder pargraphSeeder file, default 3
      */
-    public function create() {
-        //
-    }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  Request  $request
-     * @return Response
-     */
     public function store() {
+        if (!Auth::check()) {
+            return;
+        }
         $data = Paragraph::find(Input::get("id"));
         if (!$data) {
             $data = new Paragraph;
@@ -61,44 +55,10 @@ class paragraphController extends Controller {
         return $data->id;
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return Response
-     */
-    public function show($id) {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return Response
-     */
-    public function edit($id) {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  Request  $request
-     * @param  int  $id
-     * @return Response
-     */
-    public function update(Request $request, $id) {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return Response
-     */
     public function destroy() {
+        if (!Auth::check()) {
+            return;
+        }
         $data = Paragraph::find(Input::get("id"));
         $data->delete();
         return "OK";
